+++ src/Symfony/Bundle/FrameworkBundle/Tests/Fixtures/Validation/Person.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Symfony\Bundle\FrameworkBundle\Tests\Fixtures\Validation;
-
-class Person
-{
-    public $eyeColor;
-}